User Management -

+++ L.SNH.Cons/Facades/CounterFacade.cs
@@ -1,4 +1,5 @@
-﻿using L.SNH.Domain.Entities;
+﻿using L.SNH.Domain.Common;
+using L.SNH.Domain.Entities;
 using L.SNH.Domain.Repositories;
 using System;
 using System.Collections.Generic;
@@ -16,12 +17,14 @@ namespace L.SNH.Cons.Facades
         void Get();
     }
 
-    public class DefaultFacade : IDefaultFacade
+    public class CounterFacade : IDefaultFacade
     {
         protected ICountRepository CountRepo;
-        public DefaultFacade(ICountRepository _countRepo)
+        private IUnitOfWork UnitOfWork { get; set; }
+        public CounterFacade(ICountRepository _countRepo, IUnitOfWork _UnitOfWork)
         {
             this.CountRepo = _countRepo;
+            this.UnitOfWork = _UnitOfWork;
         }
 
         public void Delete()
@@ -30,8 +33,13 @@ namespace L.SNH.Cons.Facades
             CompositeCounter.Year = DateTime.Now.Year;
             CompositeCounter.Month = DateTime.Now.Month;
             CompositeCounter.CounterType = "DEF";
-            Counter Counter = this.CountRepo.GetBy(x => x.Id == CompositeCounter);
-            this.CountRepo.Delete(Counter);
+            Counter Counter = this.CountRepo.GetBy(x => x.Year == CompositeCounter.Year && x.Month == CompositeCounter.Month && x.CounterType == CompositeCounter.CounterType);
+            if (Counter!= null)
+            {
+                this.UnitOfWork.BeginTransaction();
+                this.CountRepo.Delete(Counter);
+                this.UnitOfWork.Commit();
+            }
         }
 
         public void Save()
@@ -41,7 +49,9 @@ namespace L.SNH.Cons.Facades
             CompositeCounter.Month = DateTime.Now.Month;
             CompositeCounter.CounterType = "DEF";
             Counter Counter = new Counter();
-            Counter.Id = CompositeCounter;
+            Counter.Year = CompositeCounter.Year;
+            Counter.Month = CompositeCounter.Month;
+            Counter.CounterType = CompositeCounter.CounterType;
             Counter._Counter = 1;
             Counter.CreateBy = "TEST";
             Counter.CreateDate = DateTime.Now;
@@ -51,7 +61,9 @@ namespace L.SNH.Cons.Facades
             Counter.UpdateDate = DateTime.Now;
             Counter.UpdateTerminal = "::1";
 
+            this.UnitOfWork.BeginTransaction();
             this.CountRepo.Save(Counter);
+            this.UnitOfWork.Commit();
         }
 
         public void Update()
@@ -60,9 +72,11 @@ namespace L.SNH.Cons.Facades
             CompositeCounter.Year = DateTime.Now.Year;
             CompositeCounter.Month = DateTime.Now.Month;
             CompositeCounter.CounterType = "DEF";
-            Counter Counter = this.CountRepo.GetBy(x => x.Id == CompositeCounter);
+            Counter Counter = this.CountRepo.GetBy(x => x.Year == CompositeCounter.Year && x.Month == CompositeCounter.Month && x.CounterType == CompositeCounter.CounterType);
             Counter._Counter = Counter._Counter + 212;
+            this.UnitOfWork.BeginTransaction();
             this.CountRepo.Update(Counter);
+            this.UnitOfWork.Commit();
         }
 
         public void Get()
@@ -71,7 +85,7 @@ namespace L.SNH.Cons.Facades
             CompositeCounter.Year = DateTime.Now.Year;
             CompositeCounter.Month = DateTime.Now.Month;
             CompositeCounter.CounterType = "DEF";
-            Counter Counter = this.CountRepo.GetBy(x => x.Id == CompositeCounter);
+            Counter Counter = this.CountRepo.GetBy(x => x.Year == CompositeCounter.Year && x.Month == CompositeCounter.Month && x.CounterType == CompositeCounter.CounterType);
         }
     }
 }
